Re-enable debugloadstring
Sometimes I want to run code in-game using the adonis server variable, the :terminal is laggy on my computer, and there may be other use cases for this (etc: OnStartup). It's basically the same as using :terminal with load string so there's nothing really being changed in terms of security.