The unmentionable is back.
After reviewing my previous conversation from 2018 with the admin that told me to remove said unmentionable command, I realized they said the wording was the issue. I'm assuming that means, and based on the inclusion of an identical command in other scripts, that it was the use of a specific unmentionable word that caused the issue. I've added said command back. Hopefully for good this time (I like it)