Fixed :resize
Resize, bighead, smallhead now all use ROBLOX's BodyScale/HeadScale values to handle resizing. This should ensure compatability with R15/Anthro but I don't know if it will work with R6.

Also moved NoPlayerCollide to the script's built in plugins to add a :nocollide command to disable player to player character collisions. The only reason it's a plugin is because I thought it would be better to keep all of its code together instead of adding the physics group handling stuff elsewhere in the script when this is the only thing that uses it. I might do this more going forward for the sake of organization (and might even go back and update some things to be similar to this)