Changes to client loading process
No longer using a LocalScript in ReplicatedFirst to handle loading. I suspect it randomly fails to find the client, causing issues like all client GUIs appearing on the screen (Client is in PlayerGui waiting to be loaded, any ScreenGui thats a descendant of PlayerGui gets shown regardless of how deep it is/parents)

Client loading will now be handled by the "ClientMover" localscript which will be placed in the client folder when player hooking happens. I don't expect there to be any noticable difference to users, and I'm hoping this will fix some loading errors that some users occasionally encounter (which I suspect to be related to lag/latency when loading(?))

Also the only reason it's named "ClientMover" is because it was originally just going to move the client immediately out of PlayerGui when it ran to prevent the screen covered in default guis issues, but I realized that doesn't really solve the underlying issue and instead opted to make it handle client loading from within the client loader.