Added :cross and :oldlogs
Added :cross and :oldlogs

When the game shuts down, logs will be saved to the datastore; Logs will be combined into a table with existing saved logs and sorted by time; This is intended to aid in the identification of admin abusers;

Adonis now has a system to handle cross server messaging. Functions to runs ("CrossServerCommands") can be added to Core.CrossServer table and ran (in any server) using Core.SendCrossServer(Command, Data) which will then run the command function with supplied data on every server. An example of this is the :cross command which will run any command on every server; eg. ":cross :m hi" will run ":m hi" on every running server. This command is creators only as it can be used to run ANY command on EVERY server regardless of the player's own level.

Minor console updates/changes/fixes/whatever

Also made sure the original ClientLoader is completely disabled due to a report of it kicking someone for taking too long to load (ClientMover does not have this kick, ClientLoader isn't supposed to be running at all)