Fixed !help exploit
Running "!help | !help | !help | !help ..." caused admins to be spammed with help notificiations and was bypassing the one request at a time limit. Also, because command batching runs commands in order, yielding to run each command, this created a queue of !help requests for the user where once one finished, the next one would automatically go. If you squeeze enough !helps into the command string, this could go on for a long time.

This should be good now tho.