Update WebPanel.lua
I need it to send the JobId to prevent event listeners that are still opened from closed connections. This way it doesnt try to create a web response to a request that is already closed.

I decided to just add server stats, nothing is used yet but JobId. Other stuff will be used in the next big update!