New Output & UI Fixes
Added a new output GUI to Rounded. Also fixed an issue where when a window was minimized, the dragger would be vertically thinner than it should be by 3 pixels. Especially noticable when minimizing Rounded theme windows. Also minor changes to the minimize and close buttons in the rounded theme.

Also fixed an issue with the UserPanel introduced by moving the UI folder.