"Owners" are now "HeadAdmins" (for now)
This change was done because I've always hated the Owners rank but always left it to avoid confusion. I have however decided to finally change it to something that better represents their role in Adonis.

This is subject to change pending community feedback and is not being pushed to the live release yet.

Also merged Expertcoder's additions