Update Core/Moderators/Creators
`Core`:
Changed deprecated userId and connect with the undeprecated version.

`Moderators`:
Fixed resetview if a player argument was provided. (for some reason it was empty)
Fixed a typo on the trip command.

`Core/Admin`:
Added check if the player isn't a table to kick in-game. (DirectBan fix)

`Creators`:
DirectBan used a unknown variable for the UserId inside the table (variable typo) was fixed.
Fixed globalplace Description typo.