Added Shared Folder (minor restructuring)
Added service.Yield to simplify the process of using bindableevents to yield and resume stuff

Updated Core.SavePlayerData and the clearplayerdata command to completely remove a given player's data instead of changing it to defaults.

Rewrote a lot of the queue handling stuff to be less unstable and more efficient and flexible.

Moved Service.lua to a new Server folder named "Shared"
Anything in the "Shared" folder will be cloned to client.Shared

Moved other shared stuff like credits, changelog, and FiOne there as well.

Added Core.RemoveData to handle calls to RemoveDataAsync