UI should no longer appear while loading
This was caused by the client folder, which contains all ScreenGuis related to Adonis, being placed into PlayerGui while loading. This results in a brief flash of all ScreenGuis, which are then removed by the loading process when the folder is moved to nil. To fix this, and hopefully the lag spike that sometimes comes with it, I've set Enabled to false for all ScreenGui objects in the client's UI folder. gTable:Ready() will now set Enabled to true when parenting the ScreenGui to PlayerGui when loading UI elements. Please make sure any custom themes are also setting Enabled to false for all ScreenGui objects, and are using gTable:Ready() when they are ready to be displayed to the user.