Delete ThumbnailCamera.rbxmx
Rojo is not able to correctly recreate the model's ThumbnailCamera

Every time I update the live loader, I removed the ThumbnailCamera rojo created and instead replace it with a not broken one. Currently, there is no reason to keep the ThumbnailCamera on the repo. The model used for it will be included in EXTRAS model of future releases.