Improved Hat/Face Comamnd.
`Hat`:
Uses `DonorHat` Function to give the hat. (This can prevent possible hat spam (checks if player doesn't have <15 hats) that could crash the server)

`Face`:
Checks to see if its a Image and if not it will tell the player that it isn't a valid Face.
Checks to see if its assetTypeId 18 (roblox face type) and will insert.