Fixes/Updates to Inspect
`Inspect`:
SafeChat checking fix (uses PolicyService and checks if `Discord` is able to be seen by the player, indicating that they are <13 and are unable to see social links.)
Micro-optimization to improve some performance (atleast).
Preloading to some thumbnails Inspect  adds.
Switched from `spawn` global to Adonis's `Routine` as `spawn` has a delay in some timing of execution and Routine uses coroutines, which should make it faster.