Updated ConvertPlayerCharacterToRig()
The function now supports packages as it replaces the character with a HumanoidDescription. However, custom in-game custom characters (e.g. from :char or an avatar editor) are no longer supported. (This can be fixed by updating :char to change the HumanoidDescription, which I may or may not do later. This should also fix :r15.