[Encryption/B64/GetRandom] : Improvements
**There should definitely be tests of this as it can modify and possibly break all of adonis's concurrent encryption/decryption so please make sure to stress test this.** (My tests were basic joining game and testing if :cmds and etc worked as it used Encrypt.(Client used old and Server used new and it perfectly worked))

`Client/Server | Remote`:
Applied the char var as for some reason it was never used but was set?
Switched from `string.len` to `#` as using a len using a operator is mostly faster than using a function to call to get the length and shouldn't change anything as they are the same.

`Client/Server | Functions`:
Applied some optimizations to B64Encode/Decode to not use string metamethods and to cache some global functions from being spam used to increase performance.
Applied basic caching of vars to GetRandom.