The start of basic package support
No package management or anything like that yet, and super simple. Basically the next level up from a plugin: a plugin that can add/do multiple stuff (such as client and server modules + UI or new dependencies) as opposed to a singular module or two.

Ideally I'd eventually like to have a way to load stuff like this on-demand and possibly have some sort of package management system with that.

Right now, to make a package you just create a folder and then within that folder you can mimic the hierarchy of the MainModule to merge anything you want into it. (As in, have two folders within the main folder: one for Client and one for Server, then within Client you could have a UI folder which contains UI additions, or in server a Commands folder for additional command modules, or a Plugins folder for normal plugin modules... etc etc)