Update Aliases
Introduces new aliases for commands with only one or little aliases and removes very short aliases from powerful commands (such as admin-ing someone), there are a lot of people that beg new moderators/admins to run commands such as :pa, unpa, unta, all of these are related the command, yes, but new people to adonis might now know what the alias means and we should generally avoid using short aliases, with only one letter inside that provides key meaning, in important commands.

Right now I removed these unpa, unta, ta, pa, etc. aliases, but in the future if there is a way to know which alias was used for short aliases we could also display a confirmation prompt with all of the context.