Fix chat breaking (removal of now defunct/useless specific client functions)
`Service`:
Migrate it to task library and add back tick and add task.

`Client`:
If LocalPlayer does not exist it will try to get it again. (Will halt Adonis till found)
PlayerGui uses `FindFirstChildWhichIsA`, If not found it will create a new coroutine to try and find it for over 120 Seconds, if not found it will log it as a error.

`Chat`:
Some optimizations by clamping instead of checking with `>`.