Revert #701 #702
HumanoidDescriptions have been released for a long time and should now be supported by Adonis for easier compatibility with external systems within games. If you want to be able to use image IDs for shirts or pants, you should add the old commands back in your old game with a plugin instead, as most people will not need this behavior and it is not worth sacrificing the benefits of HumanoidDescriptions just for that behavior.