Fixed ban system (#894)
* Fixed ban system.

Core:
Implement an event that will fire for any TableName that is set to allow users to know when data is being Added. (A Remove equivalent can be added by any maintainer if necessary)

Process:
Add Event check when Banned datastore is updated and has new data. This will allow for anyone who has joined first and has been DS banned to be checked and then kicked again!

For changelog (You can provide as some people may want to know of this change, its useful):
* Implemented new service.Events: DataStoreAdd_%s, This will allow you to know when certain settings are being changed.

* Change Ban check on DS set.

Instead of checking everyone for bans, it will check for the UserId of the player then kick them if detected.

* Update Process.lua

Co-authored-by: Expertcoderz <81153405+Expertcoderz@users.noreply.github.com>