Fix for the noclip argument for the ":fly" command
Well, the noclip argument is being always true whether using the ":fly" command or ":flynoclip" one. Causing the BasePart of the character to be always Nonclipped which also cause the character Not to float properly while swimming in water.
Using a conditional to determine if the "noclip" argument is actually true not just exist (a table) and if not, it will be set to false.