Re-add ChatCreateRobloxCommands (#1946)
Re-added `Settings.ChatCreateRobloxCommands`, default is currently false.
Also adds some checks to make sure HandleTextChatCommands is loaded when the setting is enabled & change timeout to 120 to prevent yielding.

Would be better for the developer to actually know the option instead of it being obscured behind source code and being on by default.